Social_Media_Bot/requirements.txt
facebook-sdk==3.1.0
tweepy==4.10.0
schedule==1.1.0